HRMNet-code/pyproject.toml
[project]
name = "hrmnetv2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "einops>=0.8.1",
    "grad-cam>=1.5.5",
    "imageio>=2.35.1",
    "ipykernel>=6.29.5",
    "matplotlib>=3.7.5",
    "omegaconf>=2.3.0",
    "opencv-python>=4.13.0.92",
    "pandas>=2.0.3",
    "pillow>=10.4.0",
    "scipy>=1.10.1",
    "thop>=0.1.1.post2209072238",
    "timm>=1.0.25",
    "torchmetrics>=1.5.2",
    "tqdm>=4.67.3",
    "yacs>=0.1.8",
]
